Include score last updated times in game state.
This makes it easier to determine the winner (the highest score, and in the event of a tie: the person who got to that score earlier.

Edge case: What if team A has 1100, team B has 1000 and A loses 100 on the very last question => Who got to 1000 earlier? Currently, we will choose team B, but technically A may have reached 1000 before B